% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_apa_table.R
\name{apa_table}
\alias{apa_table}
\title{APA-style Mplus Tables}
\usage{
apa_table(tidy_data, model_type)
}
\arguments{
\item{tidy_data}{A tidy dataset of Mplus models, created by \code{\link[=tidy_compile]{tidy_compile()}}}

\item{model_type}{One of 'null', 'univariate', or 'bivariate'.}
}
\value{
APA-style table of Mplus output.
}
\description{
Creates APA-style tables containing output from multiple Mplus objects.
}
