% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dejon_compile.R
\name{dejon_compile}
\alias{dejon_compile}
\title{Compile Mplus Data for Dejonckheere Project}
\usage{
dejon_compile(
  Mplus_file,
  model_type,
  rounding = 2,
  parameters = NULL,
  variables = NULL,
  paramheaders = NULL,
  outcomes = NULL,
  standardized = TRUE
)
}
\arguments{
\item{Mplus_file}{An Mplus object for the Dejonckheere project generated by the Mplus Automation package from Mplus output using the \code{\link[MplusAutomation:readModels]{MplusAutomation::readModels()}} function.}

\item{model_type}{One of 'null', 'univariate', or 'bivariate'.}

\item{rounding}{A value between 0 and 3. Defaults to 2.}

\item{parameters}{Parameters in the Mplus output, without the variable name at the start e.g. NAMEAN is MEAN, NASDW is SDW.
Must be in capitals, and exactly like it is in the original output. These can be found using \code{\link[=mplus_check_params]{mplus_check_params()}}.
Defaults to all available parameters.}

\item{variables}{Variables from the Mplus output. Exact variable names can be found using \code{\link[=mplus_check_params]{mplus_check_params()}}. Defaults to all available variables.}

\item{paramheaders}{Parameter headers from the Mplus output. Exact parameter headers can be found using \code{\link[=mplus_check_params]{mplus_check_params()}}. For null models,
defaults to New.Additional.Parameters. For univariate and bivariate models, defaults to Z.ON and R2.}

\item{outcomes}{Outcome variables in the Mplus output. Available outcomes can be found using \code{\link[=mplus_check_params]{mplus_check_params()}}. Defaults to all outcomes.}

\item{standardized}{Whether standardized or unstandardized output should be used for univariate and bivariate models. Defaults to TRUE.}
}
\value{
A tibble containing specified variables and parameters from multiple Mplus models.
}
\description{
Creates a tidy dataset for the Dejonckheere Project containing analyses from multiple Mplus objects. This can then be used in the \code{\link[=dejon_apa_table]{dejon_apa_table()}} function, or can be saved separately.
}
