% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dejon_tidy_data.R
\name{dejon_null}
\alias{dejon_null}
\title{Extract Variables from Null Models for Dejonckheere Project}
\usage{
dejon_null(Mplus_file, model_n = 1)
}
\arguments{
\item{Mplus_file}{An Mplus object generated by the Mplus Automation package from Mplus output using the \code{\link[MplusAutomation:readModels]{MplusAutomation::readModels()}} function. Can contain multiple MPlus files.}

\item{model_n}{If a Mplus object contains more than one model, this determines which model to select. Defaults to 1.}
}
\value{
A tibble containing the name of the dataset, paramheaders, total number of observations and participants, parameters, point estimates, CIs, and variable names.
}
\description{
Extracts specified parameters and variables from a singular Mplus null model.
}
\seealso{
\code{\link[MplusAutomation]{readModels}}
}
