% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dejon_tidy_data.R
\name{dejon_univar}
\alias{dejon_univar}
\title{Extract Variables from Univariate Models for Dejonckheere Project}
\usage{
dejon_univar(Mplus_file, model_n = 1, standardised = TRUE)
}
\arguments{
\item{Mplus_file}{An Mplus object generated by the Mplus Automation package from Mplus output using the  \code{\link[MplusAutomation]{readModels}} function. Can contain multiple MPlus files.}

\item{model_n}{If a Mplus object contains more than one model, this determines which model to select. Defaults to 1.}
}
\value{
A tibble containing the name of the dataset, paramheaders, parameters, point estimates, CIs, and names of outcome and predictor variables.
}
\description{
Extracts specified parameters and variables from a singular Mplus univariate model.
}
\seealso{
\code{\link[MplusAutomation]{readModels}}
}
