% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_file_names.R
\name{fix_file_names}
\alias{fix_file_names}
\title{Fix File Names}
\usage{
fix_file_names(dir)
}
\arguments{
\item{dir}{A directory}
}
\value{
Renamed files
}
\description{
The \code{\link[MplusAutomation:readModels]{MplusAutomation::readModels()}} function from the \code{MplusAutomation} package reads all special characters in file names, such as
brackets, dashes, and spaces, in as full stops. This is problematic for determining the name of the file. This function replaces
all special characters in file names with underscores, renaming the files. These can then be read in correctly by \code{\link[MplusAutomation:readModels]{MplusAutomation::readModels()}}.
Warning: it is recommended that you back up the files before renaming
}
