% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_apa_table.R
\name{mplus_apa_table}
\alias{mplus_apa_table}
\title{APA-style Mplus Tables}
\usage{
mplus_apa_table(
  data,
  type = "general",
  bold = TRUE,
  CI = TRUE,
  SD = FALSE,
  triangle = "lower",
  header = TRUE,
  orientation = "vertical",
  n_tables = NULL
)
}
\arguments{
\item{type}{One of 'general' or 'correlation'. Defaults to 'general'.}

\item{bold}{Whether significant results (as indicated by confidence intervals not including 0) should be bolded. Defaults to TRUE.}

\item{CI}{Whether confidence intervals should be displayed in square brackets. Defaults to TRUE.}

\item{SD}{Whether standard deviation should be displayed in round brackets. Defaults to FALSE. Note that CI and SD cannot both equal TRUE.}

\item{triangle}{Whether correlations tables, if applicable, should contain values in the upper or lower triangle. One of 'upper' or 'lower'. Defaults to "upper".}

\item{header}{Whether correlations tables should contain the name of the dataset. Defaults to TRUE.}

\item{orientation}{Whether the orientation of the table (if the type = 'general') should be vertical or horizontal. Defaults to "vertical".}
}
\value{
APA-style table of Mplus output.
}
\description{
Creates APA-style tables containing output from multiple Mplus objects. These include descriptives tables, correlations tables, multi-level reliability tables and general tables.
}
