% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_check_parameter_options.R
\name{mplus_check_params}
\alias{mplus_check_params}
\title{Check Parameters}
\usage{
mplus_check_params(
  Mplus_file,
  parameter_type,
  standardized = TRUE,
  project = "other"
)
}
\arguments{
\item{Mplus_file}{An mplus object generated by the Mplus Automation package from Mplus output using the \code{\link[MplusAutomation:readModels]{MplusAutomation::readModels()}} function.}

\item{parameter_type}{One of 'parameters', 'paramheader' or 'display'. It is also possible to select 'outcomes' and 'variables' for Dejonckheere project models.}

\item{standardized}{Whether standardized or unstandardized output should be used for univariate and bivariate models. Defaults to TRUE.}

\item{project}{Whether the parameters are for the Dejon project or another project. One of 'dejon', 'other'. Defaults to 'other'.}
}
\value{
A list of available options for \code{\link[=mplus_compile]{mplus_compile()}} or \code{\link[=dejon_compile]{dejon_compile()}}.
}
\description{
Checks the options available to select in \code{\link[=mplus_compile]{mplus_compile()}} or \code{\link[=dejon_compile]{dejon_compile()}}.
}
\seealso{
\code{\link[=mplus_compile]{mplus_compile()}} \code{\link[=dejon_compile]{dejon_compile()}} \code{\link[MplusAutomation:readModels]{MplusAutomation::readModels()}}
}
