% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEW_tidy_compile.R
\name{mplus_compile}
\alias{mplus_compile}
\title{Compile Mplus Data (NEW)}
\usage{
mplus_compile(
  Mplus_file,
  rounding = 2,
  param_header = NULL,
  parameter = NULL,
  display = "all",
  standardized = TRUE,
  converged = TRUE,
  define = FALSE
)
}
\arguments{
\item{Mplus_file}{An mplus object generated by the Mplus Automation package from Mplus output using the \code{\link[MplusAutomation:readModels]{MplusAutomation::readModels()}} function.}

\item{rounding}{a value between 0 and 3. Defaults to 2.}

\item{param_header}{parameter headers from the Mplus output. Exact parameter headers can be found using mplus_check_parameters(). For null models,
defaults to New.Additional.Parameters. For univariate and bivariate models, defaults to Z.ON and R2.}

\item{parameter}{parameters in the Mplus output, without the variable name at the start e.g. NAMEAN is MEAN, NASDW is SDW.
Must be in capitals, and exactly like it is in the original output. These can be found using mplus_check_parameters().
Defaults to all available parameters.}

\item{display}{How many columns should be displayed. Choose from "all", "minimal", "descriptives" or manually specify which columns should be displayed.}

\item{standardized}{Whether standardised or unstandardised output should be used for univariate and bivariate models. Defaults to TRUE.}

\item{converged}{If TRUE, removes non-converged models.}

\item{define}{Displays a column containing information from the 'define' input in the Mplus file. Defaults to FALSE.}
}
\value{
A tibble containing specified variables and parameters from multiple Mplus models.
}
\description{
Creates a tidy dataset containing analyses from multiple Mplus objects. This can then be used in the mplus_apa_table() function, or can be saved separately.
}
