% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_converge.R
\name{mplus_converge}
\alias{mplus_converge}
\title{Check Model Convergence}
\usage{
mplus_converge(Mplus_file)
}
\arguments{
\item{Mplus_file}{An Mplus object generated by the Mplus Automation package from Mplus output using the \code{\link[MplusAutomation]{readModels}} function. This can contain multiple models.}
}
\value{
A tidy dataset indicating whether each of the models converged.
}
\description{
Determines whether any of the Mplus models did not converge. Models that do not converge will not report confidence intervals. This is used to determine whether a model converged.
}
