% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_converge.R
\name{mplus_remove_converge}
\alias{mplus_remove_converge}
\title{Remove Non-Converged Models}
\usage{
mplus_remove_converge(Mplus_model, dir = getwd())
}
\arguments{
\item{Mplus_file}{An Mplus object generated by the Mplus Automation package from Mplus output using the \code{\link[MplusAutomation]{readModels}} function. This can contain multiple models.}
}
\value{
The original Mplus output excluding the non-converged models.
}
\description{
Checks whether any models in the list of Mplus models did not converge, and removes those that did not converge.
}
\seealso{
\code{\link[MplusReadR]{mplus_converge}}
}
