% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_tidy_data_functions.R
\name{mplus_tidy_null}
\alias{mplus_tidy_null}
\title{Extract Variables from Null Models
Extracts specified parameters and variables from a singular Mplus null model.}
\usage{
mplus_tidy_null(Mplus_file, model_n = 1)
}
\arguments{
\item{Mplus_file}{An Mplus object generated by the Mplus Automation package from Mplus output using the readModels() function. Can contain multiple MPlus files.}

\item{model_n}{If a Mplus object contains more than one model, this determines which model to select. Defaults to 1.}
}
\value{
A tibble containing the name of the dataset, paramheaders, total number of observations and participants, parameters, point estimates, CIs, and variable names.
}
\description{
Extract Variables from Null Models
Extracts specified parameters and variables from a singular Mplus null model.
}
