% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_tidy_data_functions.R
\name{tidy_bivar}
\alias{tidy_bivar}
\title{Extract Variables from Bivariate Models}
\usage{
tidy_bivar(Mplus_file, model_n = 1, standardised = TRUE)
}
\arguments{
\item{Mplus_file}{An Mplus object generated by the Mplus Automation package from Mplus output using the \code{\link[MplusAutomation]{readModels}} function. Can contain multiple MPlus files.}

\item{model_n}{If a Mplus object contains more than one model, this determines which model to select. Defaults to 1.}
}
\value{
A tibble containing the name of the dataset, paramheaders, parameters, point estimates, CIs, and names of outcome and predictor variables.
}
\description{
Extracts specified parameters and variables from a singular Mplus bivariate model.
}
\seealso{
\code{\link[MplusAutomation]{readModels}}
}
