% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_tidy_compile.R
\name{tidy_compile}
\alias{tidy_compile}
\title{Compile Mplus Data}
\usage{
tidy_compile(
  Mplus_file,
  model_type,
  rounding = 2,
  parameters = NULL,
  variables = NULL,
  paramheaders = NULL,
  outcomes = NULL,
  standardised = TRUE
)
}
\arguments{
\item{Mplus_file}{An mplus object generated by the Mplus Automation package from Mplus output using the \code{\link[MplusAutomation:readModels]{MplusAutomation::readModels()}} function.}

\item{model_type}{One of 'null', 'univariate', or 'bivariate'.}

\item{rounding}{a value between 0 and 3. Defaults to 2.}

\item{parameters}{parameters in the Mplus output, without the variable name at the start e.g. NAMEAN is MEAN, NASDW is SDW.
Must be in capitals, and exactly like it is in the original output. These can be found using mplus_check_parameters().
Defaults to all available parameters.}

\item{variables}{Variables from the Mplus output. Exact variable names can be found using mplus_check_parameters(). Defaults to all available variables.}

\item{paramheaders}{parameter headers from the Mplus output. Exact parameter headers can be found using mplus_check_parameters(). For null models,
defaults to New.Additional.Parameters. For univariate and bivariate models, defaults to Z.ON and R2.}

\item{outcomes}{Outcome variables in the Mplus output. Available outcomes can be found using mplus_check_parameters(). Defaults to all outcomes.}

\item{standardised}{Whether standardised or unstandardised output should be used for univariate and bivariate models. Defaults to TRUE.}
}
\value{
A tibble containing specified variables and parameters from multiple Mplus models.
}
\description{
Creates a tidy dataset containing analyses from multiple Mplus objects. This can then be used in the mplus_apa_table() function, or can be saved separately.
}
