% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_apa_table.R
\name{mplus_apa_table}
\alias{mplus_apa_table}
\title{APA-style Mplus Tables}
\usage{
mplus_apa_table(
  data,
  type = "general",
  bold = TRUE,
  CI = TRUE,
  SD = FALSE,
  triangle = "lower",
  header = TRUE,
  orientation = "vertical"
)
}
\value{
APA-style table of Mplus output.
}
\description{
Creates APA-style tables containing output from multiple Mplus objects. These include descriptives tables, correlations tables, multi-level reliability tables and general tables.
}
