% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_compile.R
\name{mplus_compile}
\alias{mplus_compile}
\title{Compile Mplus Data}
\usage{
mplus_compile(
  Mplus_file,
  rounding = 2,
  param_header = NULL,
  parameter = NULL,
  display = "all",
  standardized = TRUE,
  converged = TRUE,
  define = FALSE,
  dir = getwd()
)
}
\arguments{
\item{Mplus_file}{An mplus object generated by the Mplus Automation package from Mplus output using the \code{\link[MplusAutomation:readModels]{MplusAutomation::readModels()}} function.}

\item{rounding}{A value between 0 and 3. Defaults to 2.}

\item{param_header}{Parameter headers from the Mplus output. Exact parameter headers can be found using \code{\link[=mplus_check_params]{mplus_check_params()}}.}

\item{parameter}{Parameters in the Mplus output. These can be found using \code{\link[=mplus_check_params]{mplus_check_params()}}).
Defaults to all available parameters.}

\item{display}{How many columns should be displayed. Choose from "all", "minimal", "descriptives" or manually specify which columns should be displayed. Available columns can be found using \code{\link[=mplus_check_params]{mplus_check_params()}}.}

\item{standardized}{Whether standardized or unstandardized output should be used for univariate and bivariate models. Defaults to TRUE.}

\item{converged}{If TRUE, removes non-converged models.}

\item{define}{Displays a column containing information from the 'define' input in the Mplus file. Defaults to FALSE.}

\item{dir}{Directory of the data.}
}
\value{
A tibble containing specified variables and parameters from multiple Mplus models.
}
\description{
Creates a tidy dataset containing analyses from multiple Mplus objects. This can then be used in the \code{mplus_apa_table()} function, or can be saved separately.
}
